change counter to user struct

+++ programs/box_2024/src/instructions/claim_box.rs
@@ -11,7 +11,7 @@ use anchor_spl::token::Transfer;
 use crate::{BoxErrors, BoxStruct, ClaimBoxEvent, UserStruct, BOX_ACCOUNT, USER_ACCOUNT};
 
 #[derive(Accounts)]
-#[instruction(box_id: u8, id: u64)]
+#[instruction(box_id: u8, id: u16)]
 pub struct ClaimBox<'info> {
     #[account(
         // mut,
@@ -56,7 +56,7 @@ pub struct ClaimBox<'info> {
     pub system_program: Program<'info, System>,
 }
 
-pub fn claim_handler(ctx: Context<ClaimBox>, box_id: u8, id: u64) -> Result<()> {
+pub fn claim_handler(ctx: Context<ClaimBox>, box_id: u8, id: u16) -> Result<()> {
     let buyer_account = &mut ctx.accounts.buyer_account;
     let box_account = &ctx.accounts.box_account;
     // let buyer = &ctx.accounts.buyer;
